Title: 달에서 무게 구하기, Time: 94ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 250311/달에서 무게 구하기/weight-on-the-moon.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-        int weight = 13;
+        float weight = 13;
         float rate = 0.165f;
 
         System.out.print(weight + " * " + rate + " = " + weight * rate);
